dedup : CPU + Metal is working
